refactor(uc-application): route V3 codec helpers through facade
Per AGENTS.md §11.4.3, external crates should only access uc-application
through `src/facade/`. The V3 envelope codec helpers
(decode_v3_bytes_to_snapshot, decode_v3_bytes_to_snapshot_and_blob_refs,
V3BlobRef) were re-exported directly from lib.rs, letting CLI / test
consumers reach into usecases::clipboard_sync. Move the re-export under
facade/mod.rs and switch call sites (recv/send/watch + the
slice2_phase2_clipboard_e2e bootstrap test) to import via the facade
path. Internal references in app_facade.rs and clipboard_outbound/mod.rs
now reach the codec types directly through the usecases tree.
